Add compiler warnings and treat them as errors

+++ Game.cpp
@@ -49,7 +49,7 @@ bool Game::add(string playerName){
 	return true;
 }
 
-int Game::howManyPlayers()
+size_t Game::howManyPlayers()
 {
 	return players.size();
 }
